add logger to service

+++ APIorm/Services/ProdutoService.cs
@@ -6,15 +6,19 @@ using APIorm.Services.Interfaces;
 using System;
 using APIorm.Models.Interface;
 using APIorm.Exceptions;
+using Microsoft.Extensions.Logging;
 
 namespace APIorm.Services
 {
     public class ProdutoService : IProdutoService
     {
         private readonly IProdutoRepository repository;
-        public ProdutoService(IProdutoRepository repository)
+        private readonly ILogger<ProdutoService> logger;
+
+        public ProdutoService(IProdutoRepository repository, ILogger<ProdutoService> logger)
         {
             this.repository = repository;
+            this.logger = logger;
         }
 
         public async Task<ResponseCluster<IEnumerable<Produto>>> GetAll(int pageIndex, int pageSize)
@@ -32,9 +36,12 @@ namespace APIorm.Services
         public async Task<Produto> Put(long id, Produto produto)
         {
             var exist = await repository.ExistsAsync(id);
-            if(!exist)
+            if (!exist)
+            {
+                logger.LogError("ProdutoService.Put","Produto não encontrado");
                 throw new ApiException(ApiException.ApiExceptionReason.PRODUTO_NAO_ENCONTRADO, "Produto não encontrado");
-            
+            }
+
             produto.IdProduto = id;
             return await repository.UpdateAsync(produto);
         }
@@ -47,9 +54,15 @@ namespace APIorm.Services
         {
             return repository.PostList(produtoList);
         }
-        public Task<Produto> Delete(long id)
+        public async Task<Produto> Delete(long id)
         {
-            return repository.Delete(id);
+            var exist = await repository.ExistsAsync(id);
+            if (!exist)
+            {
+                logger.LogError("ProdutoService.Delete", "Produto não encontrado");
+                throw new ApiException(ApiException.ApiExceptionReason.PRODUTO_NAO_ENCONTRADO, "Produto não encontrado");
+            }
+            return await repository.Delete(id);
         }
     }
 }
